flask_app/pyproject.toml
[project]
name = "paperclip-adjacent"
version = "0.1.0"
dependencies = ["flask", "tinydb", "flask-cors", "pyyaml", "python-dotenv"]

[project.optional-dependencies]
dev = ["pytest", "requests", "nox"]

[tool.setuptools]
packages = ["api"]

[tool.pytest.ini_options]
junit_family = "xunit2"
junit_suite_name = "pytest"
junit_logging = "all"
addopts = "--junit-xml=tests/results/*.xml"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- CLI COMMANDS ---
[project.scripts]
paperclip = "api.__main__:main"